backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
pinecone-client
sentence-transformers
langchain
langchain-groq
